+++ finalware/__init__.py
@@ -1,5 +1,5 @@
 __author__ = 'Val Neekman @ Neekware Inc. [@vneekman]'
 __description__ = 'A utility application that finalizes the bootstrapping of a Django-powered site'
-__version__ = '0.0.4'
+__version__ = '0.0.5'
 
-default_app_config = 'finalware.apps.FinalwareConfig'
+default_app_config = 'finalware.apps.AppConfig'